refactor(handlers): move translateServiceError to utils.go (shared by balances + account-history)
